fix ctrl url discovery
It turns out that on yamaha receivers the discovered port (typically
port 8080) for the desc.xml *is not* actually active for the control
url, which is typically sitting on port 80 (and is expressed with the
yamaha:X_URLBase field).

This is a bit more strict about using that field to build up the
control url.

It also returns model_name which the rxv code asks for. This allows
the netdisco discovery to be a full replacement for the ssdp code in
rxv at least as far as home-assistant is concerned, and should enable
yamaha via the discovery component.