final fix for yamaha discovery (#56)
When discovering zones for the yamaha receiver, we are actually
probing the xml of the yamaha specific desc.xml, which provides all
the function details of the system (it's kind of soap-esque dump). The
generic desc.xml isn't good enough, and if we use that we don't end up
with any receivers discovered because no zones are found.