Add Bluetooth + Playstation 3 support
Adds support for Bluetooth HID devices, as well as an initial implementation
of a Bluetooth-enabled device (PS3).  The underlying Bluetooth library uses
Linux's BlueZ bluetooth stack, so currently only Linux is supported.  This
functionality works by registering the host as a Bluetooth Device.  Upon
successfully pairing the host with the Playstation 3, Home Assistant will
be able to connect to and control the PS3 via the use of HID commands.

This should also allow other Bluetooth HID-compatible systems (other gaming
consoles, Blu-Ray players, Apple TV, etc) to be accessible via Home Assistant
with code that's very similar to the Playstation 3 implementation.