Ignore bravia televisions from songpal detection
It was discovered that sony bravia televisions do also use the very same
protocol for communication as "songpal" component, causing them to be added
to homeassistant although all relevant parts for them are not yet supported.

This patch adds an ignore for those devices for the time being,
at some later point the python-songpal will likely be extended to work
more properly on those devices and this may be removed.

Related to https://github.com/home-assistant/home-assistant/issues/13022